Split general support settings search entries
